+++ service/controller/rbac/resource/fluxauth/create.go
@@ -106,12 +106,15 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	// - write-silences access for "automation" ServiceAccount *in this org namespace*
 	clusterRoleBinding := &rbacv1.ClusterRoleBinding{
 		ObjectMeta: metav1.ObjectMeta{
-			Name: pkgkey.WriteSilencesAutomationSARoleBindingName(),
+			Name: pkgkey.WriteSilencesAutomationSAinNSRoleBindingName(ns.Name),
 			Labels: map[string]string{
 				label.ManagedBy: project.Name(),
 			},
 		},
-		Subjects: []rbacv1.Subject{},
+		Subjects: []rbacv1.Subject{{
+			Kind:      "ServiceAccount",
+			Name:      pkgkey.AutomationServiceAccountName,
+			Namespace: ns.Name}},
 		RoleRef: rbacv1.RoleRef{
 			APIGroup: "rbac.authorization.k8s.io",
 			Kind:     "ClusterRole",
@@ -244,13 +247,6 @@ func (r *Resource) createOrUpdateClusterRoleBinding(ctx context.Context, ns core
 		return microerror.Mask(err)
 	} else if needsUpdateClusterRoleBinding(clusterRoleBinding, existingClusterRoleBinding) {
 		r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("updating cluster role binding %#q for Automation SA in namespace %s", clusterRoleBinding.Name, ns.Name))
-		clusterRoleBinding.Subjects = append(existingClusterRoleBinding.Subjects,
-			rbacv1.Subject{
-				Kind:      "ServiceAccount",
-				Name:      pkgkey.AutomationServiceAccountName,
-				Namespace: ns.Name,
-			},
-		)
 		_, err := r.k8sClient.RbacV1().ClusterRoleBindings().Update(ctx, clusterRoleBinding, metav1.UpdateOptions{})
 		if err != nil {
 			return microerror.Mask(err)